+++ CTFd/schemas/users.py
@@ -1,7 +1,10 @@
-from marshmallow import ValidationError, pre_load, validate
+from marshmallow import ValidationError, post_dump, pre_load, validate
+from marshmallow.fields import Nested
 from marshmallow_sqlalchemy import field_for
+from sqlalchemy.orm import load_only
 
-from CTFd.models import Users, ma
+from CTFd.models import UserFieldEntries, UserFields, Users, ma
+from CTFd.schemas.fields import UserFieldEntriesSchema
 from CTFd.utils import get_config, string_types
 from CTFd.utils.crypto import verify_password
 from CTFd.utils.email import check_email_is_whitelisted
@@ -49,6 +52,9 @@ class UserSchema(ma.ModelSchema):
     )
     country = field_for(Users, "country", validate=[validate_country_code])
     password = field_for(Users, "password")
+    fields = Nested(
+        UserFieldEntriesSchema, partial=True, many=True, attribute="field_entries"
+    )
 
     @pre_load
     def validate_name(self, data):
@@ -180,6 +186,126 @@ class UserSchema(ma.ModelSchema):
                 data.pop("password", None)
                 data.pop("confirm", None)
 
+    @pre_load
+    def validate_fields(self, data):
+        """
+        This validator is used to only allow users to update the field entry for their user.
+        It's not possible to exclude it because without the PK Marshmallow cannot load the right instance
+        """
+        fields = data.get("fields")
+        if fields is None:
+            return
+
+        current_user = get_current_user()
+
+        if is_admin():
+            user_id = data.get("id")
+            if user_id:
+                target_user = Users.query.filter_by(id=data["id"]).first()
+            else:
+                target_user = current_user
+
+            # We are editting an existing user
+            if self.view == "admin" and self.instance:
+                target_user = self.instance
+                provided_ids = []
+                for f in fields:
+                    f.pop("id", None)
+                    field_id = f.get("field_id")
+
+                    # # Check that we have an existing field for this. May be unnecessary b/c the foriegn key should enforce
+                    field = UserFields.query.filter_by(id=field_id).first_or_404()
+
+                    # Get the existing field entry if one exists
+                    entry = UserFieldEntries.query.filter_by(
+                        field_id=field.id, user_id=target_user.id
+                    ).first()
+                    if entry:
+                        f["id"] = entry.id
+                        provided_ids.append(entry.id)
+
+                # Extremely dirty hack to prevent deleting previously provided data.
+                # This needs a better soln.
+                entries = (
+                    UserFieldEntries.query.options(load_only("id"))
+                    .filter_by(user_id=target_user.id)
+                    .all()
+                )
+                for entry in entries:
+                    if entry.id not in provided_ids:
+                        fields.append({"id": entry.id})
+        else:
+            provided_ids = []
+            for f in fields:
+                # Remove any existing set
+                f.pop("id", None)
+                field_id = f.get("field_id")
+                value = f.get("value")
+
+                # # Check that we have an existing field for this. May be unnecessary b/c the foriegn key should enforce
+                field = UserFields.query.filter_by(id=field_id).first_or_404()
+
+                if field.required is True and value.strip() == "":
+                    raise ValidationError(
+                        f"Field '{field.name}' is required", field_names=["fields"]
+                    )
+
+                if field.editable is False:
+                    raise ValidationError(
+                        f"Field '{field.name}' cannot be editted",
+                        field_names=["fields"],
+                    )
+
+                # Get the existing field entry if one exists
+                entry = UserFieldEntries.query.filter_by(
+                    field_id=field.id, user_id=current_user.id
+                ).first()
+
+                if entry:
+                    f["id"] = entry.id
+                    provided_ids.append(entry.id)
+
+            # Extremely dirty hack to prevent deleting previously provided data.
+            # This needs a better soln.
+            entries = (
+                UserFieldEntries.query.options(load_only("id"))
+                .filter_by(user_id=current_user.id)
+                .all()
+            )
+            for entry in entries:
+                if entry.id not in provided_ids:
+                    fields.append({"id": entry.id})
+
+    @post_dump
+    def process_fields(self, data):
+        """
+        Handle permissions levels for fields.
+        This is post_dump to manipulate JSON instead of the raw db object
+
+        Admins can see all fields.
+        Users (self) can see their edittable and public fields
+        Public (user) can only see public fields
+        """
+        # Gather all possible fields
+        removed_field_ids = []
+        fields = UserFields.query.all()
+
+        # Select fields for removal based on current view and properties of the field
+        for field in fields:
+            if self.view == "user":
+                if field.public is False:
+                    removed_field_ids.append(field.id)
+            elif self.view == "self":
+                if field.editable is False and field.public is False:
+                    removed_field_ids.append(field.id)
+
+        # Rebuild fuilds
+        fields = data.get("fields")
+        if fields:
+            data["fields"] = [
+                field for field in fields if field["field_id"] not in removed_field_ids
+            ]
+
     views = {
         "user": [
             "website",
@@ -189,6 +315,7 @@ class UserSchema(ma.ModelSchema):
             "bracket",
             "id",
             "oauth_id",
+            "fields",
         ],
         "self": [
             "website",
@@ -200,6 +327,7 @@ class UserSchema(ma.ModelSchema):
             "id",
             "oauth_id",
             "password",
+            "fields",
         ],
         "admin": [
             "website",
@@ -217,6 +345,7 @@ class UserSchema(ma.ModelSchema):
             "password",
             "type",
             "verified",
+            "fields",
         ],
     }
 
@@ -226,5 +355,6 @@ class UserSchema(ma.ModelSchema):
                 kwargs["only"] = self.views[view]
             elif isinstance(view, list):
                 kwargs["only"] = view
+        self.view = view
 
         super(UserSchema, self).__init__(*args, **kwargs)