+++ CTFd/views.py
@@ -28,7 +28,7 @@ from CTFd.utils import config, get_config, set_config
 from CTFd.utils import user as current_user
 from CTFd.utils import validators
 from CTFd.utils.config import is_setup
-from CTFd.utils.config.pages import get_page
+from CTFd.utils.config.pages import build_html, get_page
 from CTFd.utils.config.visibility import challenges_visible
 from CTFd.utils.dates import ctf_ended, ctftime, view_after_ctf
 from CTFd.utils.decorators import authed_only
@@ -330,6 +330,30 @@ def static_html(route):
             return redirect(url_for("auth.login", next=request.full_path))
 
         return render_template("page.html", content=page.content)
+
+
+@views.route("/tos")
+def tos():
+    tos_url = get_config("tos_url")
+    tos_text = get_config("tos_text")
+    if tos_url:
+        return redirect(tos_url)
+    elif tos_text:
+        return render_template("page.html", content=build_html(tos_text))
+    else:
+        abort(404)
+
+
+@views.route("/privacy")
+def privacy():
+    privacy_url = get_config("privacy_url")
+    privacy_text = get_config("privacy_text")
+    if privacy_url:
+        return redirect(privacy_url)
+    elif privacy_text:
+        return render_template("page.html", content=build_html(privacy_text))
+    else:
+        abort(404)
 
 
 @views.route("/files", defaults={"path": ""})